auth-service/requirements.txt
django==1.11
django-rest-framework
djangorestframework-simplejwt
django-cas-ng
coreapi

